backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
redis==5.0.1
django-cors-headers==4.3.1
pytest==8.2.1
pytest-django==4.8.0
python-dotenv==1.0.1
